Make spa much more standalone
