+++ src/argus/site/settings/test_CI.py
@@ -6,7 +6,7 @@ from django.utils.log import DEFAULT_LOGGING
 
 from . import get_bool_env, get_str_env
 from .base import *  # noqa: F403
-from argus.spa.settings import *  # noqa: F403
+from argus.spa.spa_settings import *  # noqa: F403
 
 
 DEBUG = get_bool_env("DEBUG", True)
